AnalogSource: return a float from read() method
this gives the full resolution when using the ADC backend. It also
adds scaling for the Arduino backend

+++ libraries/AP_AnalogSource/AnalogSource.h
@@ -5,7 +5,7 @@
 class AP_AnalogSource
 {
     public:
-    virtual int read(void) = 0;
+    virtual float read(void) = 0;
 };
 
 #endif // __ANALOG_SOURCE_H__
